Hide individual theos inside create_indiv
Those are only an implementation detail.
